renamed standarderror.rb --> StandardError.rb
