+++ lib/bitstampede/StandardError.rb
